Added placeholder pages for additional content
Because
- I needed places to start building additional pages
- Added a top menu to access those additional pages
- Attempted to put the footer at the very bottom of page

I added links to the blog, projects, and resume_CV pages. I also created those
HTML files. I also attempted to make the footer at the bottom of the page but
only succeeded in pushing the footer down. I want the footer to be at the bottom
of the page regardless of whether the content does not fill the page, but I also
do not want the footer to cover content if increased content is added to the page.