Moved main css out of index.html style tag and to own stylesheet
Because
- It seems in reading that this is best practice
- In order to start using the same stylesheet with other pages
- In order to have a common look amongst all the pages

I created a 'sheets' folder in order to maintain .js and .css files.
This is located within the 'assets' folder as that made the most sense.
I then updated the index.html file to link to the new main.css file. The goal
is to make this file a base file to use with all the other pages on the
website and link them aesthetically.