Created copy of Google Project in projects page
Because
- Fixes #2 with 404 error from google projects

I could not link from the the_odin_project repository where the original
file was stored. Since it was a single file just moved it to the projects
folder and renamed it from index.html to google.html. I then corrected the
link in the iFrame that was creating the thumbnail on the projects.html page.