Moved all CSS to individual files, normalized CSS
Because
- It keeps different languages in their own files
- Allows use across different pages for project consistency

Got the normalize.css file from github.com/necolas/normalize.css. The
purpose is to normalize CSS and HTML elements across the more popular
internet browsers. I then created a main.css file that is shared among
all the pages in order to have a file that will keep each of the pages
consistent. Finally, I moved the current styling to individual CSS files
that can then be updated. Specifically for the blog and projects pages.